requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.22.0
streamlit==1.37.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.11.7
httpx==0.24.1
python-dotenv==1.0.1
deep-translator==1.11.4
sqlalchemy==2.0.43
pandas==2.3.2
python-multipart==0.0.20
bcrypt==4.3.0
